Add files via upload
As the project evolves, this architecture may change to accommodate additional datasets, technologies, or processes. This solution architecture is designed to be flexible and scalable. Changes will be tracked through GitHub commits and reflected in the documentation's version history. So change the file type to README.md. We will continuously optimize the architecture as the project evolves to meet new requirements.